# openapi/openapi.json
{
  "components" : {
    "schemas" : {
...
      "ActivityHeatmapBucketThresholdResponse" : {
        "properties" : {
          "bucket" : {
            "format" : "int32",
            "type" : "integer"
          },
          "maxEntryCount" : {
            "format" : "int32",
            "type" : "integer"
          },
          "minEntryCount" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "ActivityHeatmapDayResponse" : {
        "properties" : {
          "date" : {
            "format" : "date",
            "type" : "string"
          },
          "entryCount" : {
            "format" : "int32",
            "type" : "integer"
          },
          "finalizedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "intensity" : {
            "format" : "int32",
            "type" : "integer"
          },
          "logged" : {
            "type" : "boolean"
          },
          "score" : {
            "format" : "int32",
            "type" : "integer"
          },
          "scoreBand" : {
            "enum" : [ "EXCELLENT", "GOOD", "FAIR", "POOR" ],
            "type" : "string"
          },
          "scoreMode" : {
            "enum" : [ "GOAL_ADHERENCE", "CONSISTENCY" ],
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "ActivityHeatmapResponse" : {
        "properties" : {
          "bucketThresholds" : {
            "items" : {
              "$ref" : "#/components/schemas/ActivityHeatmapBucketThresholdResponse"
            },
            "type" : "array"
          },
          "days" : {
            "items" : {
              "$ref" : "#/components/schemas/ActivityHeatmapDayResponse"
            },
            "type" : "array"
          },
          "from" : {
            "format" : "date",
            "type" : "string"
          },
          "locale" : {
            "type" : "string"
          },
          "maxEntryCount" : {
            "format" : "int32",
            "type" : "integer"
          },
          "timezone" : {
            "type" : "string"
          },
          "to" : {
            "format" : "date",
            "type" : "string"
          },
          "totalLoggedDays" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "AdminSubscriptionPriceResponse" : {
        "properties" : {
          "activationConflictReason" : {
            "type" : "string"
          },
          "activationConflictedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "active" : {
            "type" : "boolean"
          },
          "amount" : {
            "type" : "number"
          },
          "badge" : {
            "type" : "string"
          },
          "baseAmount" : {
            "type" : "number"
          },
          "billingPeriodDays" : {
            "format" : "int32",
            "type" : "integer"
          },
          "createdBy" : {
            "type" : "string"
          },
          "currency" : {
            "type" : "string"
          },
          "deactivatedBy" : {
            "type" : "string"
          },
          "deactivationReason" : {
            "type" : "string"
          },
          "discountPercent" : {
            "type" : "number"
          },
          "expectedPredecessorId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "internalNotes" : {
            "type" : "string"
          },
          "planId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "scheduled" : {
            "type" : "boolean"
          },
          "validFrom" : {
            "format" : "date-time",
            "type" : "string"
          },
          "validUntil" : {
            "format" : "date-time",
            "type" : "string"
          },
          "version" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "AffiliateResponse" : {
        "properties" : {
          "applicablePlanId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "applicablePriceId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "code" : {
            "type" : "string"
          },
          "commissionPercentage" : {
            "type" : "number"
          },
          "customerPaidAmount" : {
            "type" : "number"
          },
          "discountPercentage" : {
            "type" : "number"
          },
          "displayName" : {
            "type" : "string"
          },
          "earnedAmount" : {
            "type" : "number"
          },
          "endsAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "internalNotes" : {
            "type" : "string"
          },
          "linkedUserId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "maxRedemptions" : {
            "format" : "int32",
            "type" : "integer"
          },
          "startsAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "status" : {
            "enum" : [ "ACTIVE", "INACTIVE" ],
            "type" : "string"
          },
          "successfulCustomers" : {
            "format" : "int64",
            "type" : "integer"
          },
          "version" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "DeactivateSubscriptionPriceRequest" : {
        "properties" : {
          "reason" : {
            "minLength" : 1,
            "type" : "string"
          }
        },
        "required" : [ "reason" ],
        "type" : "object"
      },
...
      "MeResponse" : {
        "properties" : {
          "calculatorRerunPromptAcknowledgedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "displayName" : {
            "type" : "string"
          },
          "email" : {
            "type" : "string"
          },
          "emailVerificationStatus" : {
            "type" : "string"
          },
          "hasPassword" : {
            "type" : "boolean"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "locale" : {
            "type" : "string"
          },
          "onboardingWelcomeSeenAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "phoneNumber" : {
            "type" : "string"
          },
          "phoneVerificationStatus" : {
            "type" : "string"
          },
          "role" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "timezone" : {
            "type" : "string"
          },
          "updatedAt" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "UpdateMeProfileRequest" : {
        "properties" : {
          "displayName" : {
            "maxLength" : 120,
            "minLength" : 0,
            "type" : "string"
          },
          "locale" : {
            "maxLength" : 16,
            "minLength" : 0,
            "type" : "string"
          },
          "timezone" : {
            "maxLength" : 64,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    }
  },
  "info" : {
    "description" : "Public HTTP contract for Gyro clients.",
    "title" : "Gyro API",
    "version" : "1.0.0"
  },
  "openapi" : "3.1.0",
  "paths" : {
...
    "/api/v1/admin/affiliates/{id}/deactivate" : {
      "post" : {
        "operationId" : "deactivate_1",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "uuid",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AffiliateResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "admin-affiliate-controller" ]
      }
    },
...
    "/api/v1/admin/subscription-prices/{id}/deactivate" : {
      "post" : {
        "operationId" : "deactivate",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DeactivateSubscriptionPriceRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AdminSubscriptionPriceResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "admin-subscription-price-controller" ]
      }
    },
...
    "/api/v1/users/me/activity-heatmap" : {
      "get" : {
        "operationId" : "activityHeatmap",
        "parameters" : [ {
          "in" : "query",
          "name" : "from",
          "required" : true,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "to",
          "required" : true,
          "schema" : {
            "format" : "date",
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ActivityHeatmapResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "me-controller" ]
      }
    },
...
    "/api/v1/users/me/profile" : {
      "patch" : {
        "operationId" : "updateProfile",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateMeProfileRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/MeResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [ "me-controller" ]
      }
    },
...
  },
  "servers" : [ {
    "url" : "/"
  } ]
}